+++ driver/src/main/java/org/kaazing/k3po/driver/internal/behavior/handler/command/AbortHandler.java
@@ -1,0 +1,38 @@
+/**
+ * Copyright 2007-2015, Kaazing Corporation. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kaazing.k3po.driver.internal.behavior.handler.command;
+
+import static org.kaazing.k3po.driver.internal.netty.channel.Channels.abort;
+
+import org.jboss.netty.channel.ChannelFuture;
+import org.jboss.netty.channel.ChannelHandlerContext;
+
+public class AbortHandler extends AbstractCommandHandler {
+
+    @Override
+    protected void invokeCommand(ChannelHandlerContext ctx) throws Exception {
+
+        ChannelFuture handlerFuture = getHandlerFuture();
+        abort(ctx, handlerFuture);
+    }
+
+    @Override
+    protected StringBuilder describe(StringBuilder sb) {
+        return sb.append("abort");
+    }
+
+}