+++ lang/src/test/java/org/kaazing/k3po/lang/internal/parser/ScriptParserImplTest.java
@@ -15,6 +15,9 @@
  */
 package org.kaazing.k3po.lang.internal.parser;
 
+import static org.junit.Assert.assertEquals;
+import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.ABORT;
+import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.ABORTED;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.ACCEPT;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.CLOSE;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.CLOSED;
@@ -27,14 +30,14 @@ import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.LITERAL_
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.PROPERTY_NODE;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.READ;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.READ_AWAIT;
-import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.READ_NOTIFY;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.READ_MASK_OPTION;
+import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.READ_NOTIFY;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.SCRIPT;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.VARIABLE_LENGTH_BYTES_MATCHER;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.WRITE;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.WRITE_AWAIT;
-import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.WRITE_NOTIFY;
 import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.WRITE_MASK_OPTION;
+import static org.kaazing.k3po.lang.internal.parser.ScriptParseStrategy.WRITE_NOTIFY;
 import static org.kaazing.k3po.lang.internal.regex.NamedGroupPattern.compile;
 import static org.kaazing.k3po.lang.internal.test.junit.Assert.assertEquals;
 
@@ -47,6 +50,8 @@ import javax.el.ValueExpression;
 
 import org.junit.Ignore;
 import org.junit.Test;
+import org.kaazing.k3po.lang.internal.ast.AstAbortNode;
+import org.kaazing.k3po.lang.internal.ast.AstAbortedNode;
 import org.kaazing.k3po.lang.internal.ast.AstAcceptNode;
 import org.kaazing.k3po.lang.internal.ast.AstCloseNode;
 import org.kaazing.k3po.lang.internal.ast.AstClosedNode;
@@ -61,6 +66,8 @@ import org.kaazing.k3po.lang.internal.ast.AstWriteAwaitNode;
 import org.kaazing.k3po.lang.internal.ast.AstWriteNotifyNode;
 import org.kaazing.k3po.lang.internal.ast.AstWriteOptionNode;
 import org.kaazing.k3po.lang.internal.ast.AstWriteValueNode;
+import org.kaazing.k3po.lang.internal.ast.builder.AstAbortNodeBuilder;
+import org.kaazing.k3po.lang.internal.ast.builder.AstAbortedNodeBuilder;
 import org.kaazing.k3po.lang.internal.ast.builder.AstAcceptNodeBuilder;
 import org.kaazing.k3po.lang.internal.ast.builder.AstCloseNodeBuilder;
 import org.kaazing.k3po.lang.internal.ast.builder.AstClosedNodeBuilder;
@@ -835,6 +842,32 @@ public class ScriptParserImplTest {
         AstCloseNode actual = parser.parseWithStrategy(scriptFragment, CLOSE);
 
         AstCloseNode expected = new AstCloseNodeBuilder().done();
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void shouldParseAbort() throws Exception {
+
+        String scriptFragment = "abort";
+
+        ScriptParserImpl parser = new ScriptParserImpl();
+        AstAbortNode actual = parser.parseWithStrategy(scriptFragment, ABORT);
+
+        AstAbortNode expected = new AstAbortNodeBuilder().done();
+
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    public void shouldParseAborted() throws Exception {
+
+        String scriptFragment = "aborted";
+
+        ScriptParserImpl parser = new ScriptParserImpl();
+        AstAbortedNode actual = parser.parseWithStrategy(scriptFragment, ABORTED);
+
+        AstAbortedNode expected = new AstAbortedNodeBuilder().done();
 
         assertEquals(expected, actual);
     }